// frontend/src/api/model/githubCom4H1RZooraInternalDomainNegativeMarkConfig.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainNegativeMarkMode } from './githubCom4H1RZooraInternalDomainNegativeMarkMode';

export interface GithubCom4H1RZooraInternalDomainNegativeMarkConfig {
  fraction?: number;
  mode?: GithubCom4H1RZooraInternalDomainNegativeMarkMode;
  negative_value?: number;
  wrongs_per_point?: number;
}