// frontend/src/api/model/githubCom4H1RZooraInternalDomainQuizRoom.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainClassSession } from './githubCom4H1RZooraInternalDomainClassSession';
import type { GithubCom4H1RZooraInternalDomainQuiz } from './githubCom4H1RZooraInternalDomainQuiz';

export interface GithubCom4H1RZooraInternalDomainQuizRoom {
  class_session?: GithubCom4H1RZooraInternalDomainClassSession;
  class_session_id?: string;
  created_at?: string;
  ended_at?: string;
  id?: string;
  quiz?: GithubCom4H1RZooraInternalDomainQuiz;
  quiz_id?: string;
  started_at?: string;
  updated_at?: string;
}